rearrange types within files
